Fixed early mapbox inizialization
