+++ mapbox/src/main/java/org/odk/collect/mapbox/MapboxMapFragment.kt
@@ -12,7 +12,6 @@ import androidx.appcompat.content.res.AppCompatResources
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.lifecycleScope
-import androidx.startup.AppInitializer
 import com.google.android.gms.location.LocationListener
 import com.mapbox.android.core.location.LocationEngineProvider
 import com.mapbox.android.core.location.LocationEngineRequest
@@ -32,7 +31,6 @@ import com.mapbox.maps.extension.style.sources.addSource
 import com.mapbox.maps.extension.style.sources.generated.RasterSource
 import com.mapbox.maps.extension.style.sources.generated.VectorSource
 import com.mapbox.maps.extension.style.sources.getSource
-import com.mapbox.maps.loader.MapboxMapsInitializer
 import com.mapbox.maps.plugin.LocationPuck2D
 import com.mapbox.maps.plugin.animation.MapAnimationOptions.Companion.mapAnimationOptions
 import com.mapbox.maps.plugin.animation.flyTo
@@ -50,7 +48,6 @@ import com.mapbox.maps.plugin.locationcomponent.location
 import com.mapbox.maps.plugin.scalebar.scalebar
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import org.odk.collect.androidshared.data.getState
 import org.odk.collect.androidshared.utils.ScreenUtils
 import org.odk.collect.maps.MapFragment
 import org.odk.collect.maps.MapFragment.ErrorListener
@@ -183,7 +180,6 @@ class MapboxMapFragment :
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        initializeMapbox(context)
 
         val configurator = MapboxMapConfigurator()
 
@@ -192,17 +188,6 @@ class MapboxMapFragment :
             settingsProvider.getUnprotectedSettings(),
             this::onConfigChanged
         )
-    }
-
-    private fun initializeMapbox(context: Context) {
-        // Auto initialization is disabled to stop x86 devices from crashing so we do this manually
-        context.getState().let {
-            if (!it.get(KEY_MAPBOX_INITIALIZED, false)) {
-                AppInitializer.getInstance(context)
-                    .initializeComponent(MapboxMapsInitializer::class.java)
-                it.set(KEY_MAPBOX_INITIALIZED, true)
-            }
-        }
     }
 
     override fun onStart() {
@@ -609,7 +594,5 @@ class MapboxMapFragment :
 
     companion object {
         const val KEY_STYLE_URL = "STYLE_URL"
-
-        private const val KEY_MAPBOX_INITIALIZED = "mapbox_initialized"
     }
 }
